kraken api cleanup

+++ iTest/sendEmail.py
@@ -25,7 +25,7 @@ def sendEmail():
 def sendEmailBTC():
     heplers.testTitle("Sending email")
     TEST_JSON = heplers.createJson(
-        position='short',
+        position='long',
         pair="XXBTZUSD",
         size=0.01,
         stopLossCanldes=4,
